add some array bounds check optimizations

+++ src/EventStore.ClientAPI/StreamMetadataExtensions.cs
@@ -19,12 +19,12 @@ namespace EventStore.ClientAPI
       if (!string.IsNullOrWhiteSpace(metadata.CustomMetadata))
       {
         String[] ss = metadata.CustomMetadata.Split(s_separators, StringSplitOptions.RemoveEmptyEntries);
-        if (ss != null || ss.Length > 0)
+        if (ss != null || (uint)ss.Length > 0u)
         {
           foreach (var item in ss)
           {
             var kvs = item.Split(s_nameValueSeparator, StringSplitOptions.RemoveEmptyEntries);
-            if (kvs != null && kvs.Length > 0)
+            if (kvs != null && (uint)kvs.Length > 0u)
             {
               if (kvs.Length == 2)
               {
