+++ src/EventStore.Common/Utils/HostName.cs
@@ -24,7 +24,7 @@ namespace EventStore.Common.Utils
         {
             //TODO: encode???
             var path = string.Format(relativeUri, arg);
-            if(path.Length > 0 && path[0] == '/') path = path.Substring(1);
+            if((uint)path.Length > 0u && path[0] == '/') path = path.Substring(1);
             return new UriBuilder(responseUrl.Scheme, responseUrl.Host, responseUrl.Port, responseUrl.LocalPath + path).Uri.AbsoluteUri;
         }
     }
