+++ src/EventStore.Core/Services/PersistentSubscription/PersistentSubscription.cs
@@ -146,7 +146,7 @@ namespace EventStore.Core.Services.PersistentSubscription
                 {
                     _streamBuffer.AddReadMessage(new OutstandingMessage(ev.OriginalEvent.EventId, null, ev, 0));
                 }
-                if (events.Length > 0)
+                if ((uint)events.Length > 0u)
                 {
                     _statistics.SetLastKnownEventNumber(events[events.Length - 1].OriginalEventNumber);
                 }
